+++ app/src/main/java/ru/mavrinvladislav/shifttask2025/authorization/data/remote/source/AuthorizationRemoteDataSource.kt
@@ -1,10 +1,5 @@
 package ru.mavrinvladislav.shifttask2025.authorization.data.remote.source
 
-import ru.mavrinvladislav.shifttask2025.authorization.data.remote.dto.CreateOtpRequest
-import ru.mavrinvladislav.shifttask2025.authorization.data.remote.dto.CreateOtpResponse
-import ru.mavrinvladislav.shifttask2025.authorization.data.remote.dto.SignInResponse
-import ru.mavrinvladislav.shifttask2025.core.common.remote.ServerResponse
-
 interface AuthorizationRemoteDataSource {
 
     suspend fun signIn(
@@ -14,5 +9,5 @@ interface AuthorizationRemoteDataSource {
 
     suspend fun createOtp(
         phoneNumber: String
-    ): Double
+    ): Long
 }
